+++ pyproject.toml
@@ -1,9 +1,9 @@
 [project]
-name = "karp-resource-fula-ordboken"
+name = "resource-fula-ordboken"
 version = "0.1.0"
 description = "Default template for PDM package"
 authors = [
-    {name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se"},
+    { name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se" },
 ]
 dependencies = [
     "karp-lex-core>=0.6.2",
@@ -13,7 +13,7 @@ dependencies = [
 ]
 requires-python = ">=3.10"
 readme = "README.md"
-license = {text = "MIT"}
+license = { text = "MIT" }
 
 [build-system]
 requires = ["pdm-backend"]
@@ -24,9 +24,4 @@ build-backend = "pdm.backend"
 distribution = true
 
 [tool.pdm.dev-dependencies]
-dev = [
-    "ruff>=0.4.1",
-    "mypy>=1.10.0",
-    "pytest>=8.1.1",
-    "pytest-cov>=5.0.0",
-]
+dev = ["ruff>=0.4.1", "mypy>=1.10.0", "pytest>=8.1.1", "pytest-cov>=5.0.0"]
