+++ tests/test_version.py
@@ -1,5 +1,5 @@
-import karp_resource_fula_ordboken
+import resource_fula_ordboken
 
 
 def test_version():
-    assert karp_resource_fula_ordboken.get_version() == "0.1.0"
+    assert resource_fula_ordboken.get_version() == "0.1.0"
